+++ release-notes.md
@@ -4,6 +4,7 @@ RELEASE NOTES
 - add isDisplayed method in WebLocator
 - fix properties.load from inputStream correctly
 - add template for title
+- add try and catch for WebDriverException at doClick and submit
 
 **Release Notes for Testy 1.8.1**
 - add isDisplayed, isEnabled and submit method in WebLocatorDriverExecutor